Fix integer dtype

+++ tests/dsp/multirate/test_interpolator.py
@@ -19,7 +19,7 @@ def test_non_streaming_rate():
     N = 50
     rng = np.random.default_rng()
     x = rng.standard_normal(N) + 1j * rng.standard_normal(N)  # Input signal
-    r = rng.integers(3, 7)  # Interpolation rate
+    r = rng.integers(3, 7, dtype=int)  # Interpolation rate
 
     fir = sdr.Interpolator(r)
     y = fir(x, mode)
@@ -37,7 +37,7 @@ def test_non_streaming_full():
     N = 50
     rng = np.random.default_rng()
     x = rng.standard_normal(N) + 1j * rng.standard_normal(N)  # Input signal
-    r = rng.integers(3, 7)  # Interpolation rate
+    r = rng.integers(3, 7, dtype=int)  # Interpolation rate
 
     fir = sdr.Interpolator(r)
     y = fir(x, mode)
@@ -55,7 +55,7 @@ def test_streaming():
     N = 50
     rng = np.random.default_rng()
     x = rng.standard_normal(N) + 1j * rng.standard_normal(N)  # Input signal
-    r = rng.integers(3, 7)  # Interpolation rate
+    r = rng.integers(3, 7, dtype=int)  # Interpolation rate
 
     fir = sdr.Interpolator(r, streaming=True)
 
@@ -76,7 +76,7 @@ def test_streaming_match_full():
     N = 50
     rng = np.random.default_rng()
     x = rng.standard_normal(N) + 1j * rng.standard_normal(N)  # Input signal
-    r = rng.integers(3, 7)  # Interpolation rate
+    r = rng.integers(3, 7, dtype=int)  # Interpolation rate
 
     fir1 = sdr.Interpolator(r)
     y_full = fir1(x, mode="full")
